Score send fixed, bullet shooting removed for simplicity

+++ Assets/GameManager/GameManager.cs
@@ -117,7 +117,7 @@ public class GameManager : Singleton<GameManager>
     {
         if (Server != null)
         {
-            Server.GameOver(winner.ID, second.ID, winner.ScoreValue.Value);
+            Server.GameOver(winner.ID, second.ID, (int)winner.ScoreValue.Value);
             return;
         }
